[STABLEHLO] Move utils.h to include/
